Update rootRouter handling
