+++ packages/api/lib/distribution.js
@@ -2,9 +2,12 @@ const isEmpty = require('lodash/isEmpty');
 const { getSecretString } = require('@cumulus/aws-client/SecretsManager');
 const { CognitoClient, EarthdataLoginClient } = require('@cumulus/oauth-client');
 const { s3 } = require('@cumulus/aws-client/services');
+const { randomId } = require('@cumulus/common/test-utils');
+const { RecordDoesNotExist } = require('@cumulus/errors');
 
 const { isLocalApi } = require('./testUtils');
 const { AccessToken } = require('../models');
+const { isAccessTokenExpired } = require('./token');
 
 // Running API locally will be on http, not https, so cookies
 // should not be set to secure for local runs of the API.
@@ -83,8 +86,88 @@ function buildErrorTemplateVars(query) {
   return templateVars;
 }
 
+/**
+ * Helper function to pull bucket out of a path string.
+ * Will ignore leading slash.
+ * "/bucket/key" -> "bucket"
+ * "bucket/key" -> "bucket"
+ *
+ * @param {string} path - express request path parameter
+ * @returns {string} the first part of a path which is our bucket name
+ */
+function bucketNameFromPath(path) {
+  return path.split('/').filter((d) => d).shift();
+}
+
+/**
+ * Reads the input path and determines if this is a request for public data
+ * or not.
+ *
+ * @param {string} path - req.path paramater
+ * @returns {boolean} - whether this request goes to a public bucket
+ */
+function isPublicRequest(path) {
+  try {
+    const publicBuckets = process.env.public_buckets.split(',');
+    const requestedBucket = bucketNameFromPath(path);
+    return publicBuckets.includes(requestedBucket);
+  } catch (error) {
+    return false;
+  }
+}
+
+/**
+ * Ensure request is authorized through EarthdataLogin or redirect to become so.
+ *
+ * @param {Object} req - express request object
+ * @param {Object} res - express response object
+ * @param {Function} next - express middleware callback function
+ * @returns {Promise<Object>} - promise of an express response object
+ */
+async function ensureAuthorizedOrRedirect(req, res, next) {
+  // Skip authentication for debugging purposes.
+  if (process.env.FAKE_AUTH) {
+    req.authorizedMetadata = { userName: randomId('username') };
+    return next();
+  }
+
+  if (isPublicRequest(req.path)) {
+    req.authorizedMetadata = { userName: 'unauthenticated user' };
+    return next();
+  }
+
+  const {
+    accessTokenModel,
+    oauthClient,
+  } = await getConfigurations();
+
+  const redirectURLForAuthorizationCode = oauthClient.getAuthorizationUrl(req.path);
+  const accessToken = req.cookies.accessToken;
+
+  if (!accessToken) return res.redirect(307, redirectURLForAuthorizationCode);
+
+  let accessTokenRecord;
+  try {
+    accessTokenRecord = await accessTokenModel.get({ accessToken });
+  } catch (error) {
+    if (error instanceof RecordDoesNotExist) {
+      return res.redirect(307, redirectURLForAuthorizationCode);
+    }
+
+    throw error;
+  }
+
+  if (isAccessTokenExpired(accessTokenRecord)) {
+    return res.redirect(307, redirectURLForAuthorizationCode);
+  }
+
+  req.authorizedMetadata = { userName: accessTokenRecord.username };
+  return next();
+}
+
 module.exports = {
   buildErrorTemplateVars,
+  ensureAuthorizedOrRedirect,
   getConfigurations,
   useSecureCookies,
 };
